added user interface
