+++ package.json
@@ -9,9 +9,9 @@
     "express": "^4.17.3",
     "nodemon": "^2.0.15"
   },
-  "devDependencies": {},
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon app.js"
   },
   "author": "",
   "license": "ISC"
