--- views/mainpage.ejs
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>main_page</title>
-  
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-<script src="https://kit.fontawesome.com/a076d05399.js"
-crossorigin="anonymous"></script>
-<link rel="stylesheet"
-href="https://fonts.googleapis.com/css?family=Tangerine">
-  <link rel="stylesheet" href="css/mvp.css" />
-  
-</head>
-
-<style>
-*{box-sizing: border-box;}
-.boxes{
-border: 1px solid;
-float:left;
-width: 50%;
-
-
-  }
-@media (min-width: 768px) {
-  
-  
-}
-</style>
-   <body style="background-color:black;">
-  
-    <nav><img src="https://upload.wikimedia.org/wikipedia/en/f/f7/M.B.M._Engineering_College_-_Logo.png" style="width:42px;height:44px;object-fit:cover;object-position:60% 60%">
-    
-    <lable class ="logo" >E-VOTING</lable>
-    
-    <h1 style="color:white;"><center>MAIN_1st_Screen</center></h1>
-  
-  </nav>
-
-   <main> 
-   
-
-   <section class="boxes" >
-    <form >
-      <p><a style="color:blue"; href="/SignIn" >ALL EVENTS</a></p> </p>If not Registered</a></p>
-      
-      <div class = "row">
-      </form>
-    </section>
-        
-          <section class="boxes" >
-            <form >
-              <p><a style="color:blue"; href="/SignUp">SIGN UP</a></p>If not Registered</a></p>
-              
-              <div class = "row">
-              </form>
-             
-            </section>
-     
-   
-    
-</body>
-</html>
