+++ views/partials/footer.ejs
@@ -1,0 +1,8 @@
+</div>
+<div class="footer-padding"></div>
+  <div class="footer">
+    <p>©️ 2022 Copyright: E-voting</p>
+  </div>
+</div>
+</body>
+</html>
