+++ views/partials/head.ejs
@@ -1,0 +1,8 @@
+<meta charset="UTF-8">
+<title>E-vote</title>
+
+<!-- CSS (load bootstrap from a CDN)  -->
+ <!-- <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.5.2/css/bootstrap.min.css"> -->
+<!-- <style>
+  body { padding-top:50px; }
+</style> -->
